requirements.txt
anthropic>=0.40.0
google-generativeai>=0.8.0
notebooklm-py>=0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
playwright>=1.40.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
yt-dlp>=2024.1.0
feedparser>=6.0.0
